backend/translator/requirements.txt
Flask
Flask-SocketIO
google-cloud-speech
google-cloud-translate
